Build out actual logic fo creating a subscription
This is a very crude implementation of pretty much all aspects outside
of creating the actual token.

This implementation obviously has it's limits. If a card is provided
then it will create a new card token for that customer whether they want
to use that card in the future or not.